20231204 將錢包金額Int改為Double

+++ DefiAPP/Tabar/Purchase/Model/PurchaseModel.swift
@@ -17,8 +17,8 @@ struct ProductData: Codable {
     let asset: String?
     let chain: String?
     let lDs: Int?
-    let maximumAmount: Int?
-    let minumumAmount: Int?
+    let maximumAmount: Double?
+    let minumumAmount: Double?
     let title: String?
     let content: String?
     let period: Period
